@@ -24,7 +24,8 @@ namespace API.Endpoints.Customers
                })
                .WithTags("Customers")
                .Produces<ICommandResult<CustomerOutputDto?>>()
-               .WithName("Customer.GetByCpf");
+               .WithName("Customer.GetByCpf")
+               .RequireAuthorization();
         }
     }
 }
