@@ -2,6 +2,7 @@
 using Application.Interfaces.DataSources;
 using Infrastructure.DataSources;
 using Infrastructure.DbContexts;
+using Microsoft.AspNetCore.Authorization;
 using MiniValidation;
 using Shared.DTO.Ingredient.Output;
 using Shared.Result;
@@ -24,7 +25,7 @@ internal sealed class GetAll : IEndpoint
            })
            .WithTags("Ingredients")
            .Produces<ICommandResult<List<IngredientOutputDto>>>()
-           .WithName("Ingredient.GetAll");
+           .WithName("Ingredient.GetAll").RequireAuthorization();
     }
 }
 
