@@ -2,6 +2,7 @@
 using Application.Interfaces.DataSources;
 using Infrastructure.DataSources;
 using Infrastructure.DbContexts;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using MiniValidation;
 using Shared.DTO.Ingredient.Output;
@@ -28,6 +29,6 @@ internal sealed class Update : IEndpoint
            })
            .WithTags("Ingredients")
            .Produces<ICommandResult<IngredientOutputDto?>>()
-           .WithName("Ingredient.Update");
+           .WithName("Ingredient.Update").RequireAuthorization(new AuthorizeAttribute { Roles = "Admin,Master" });
     }
 }
