@@ -5,6 +5,7 @@ using Infrastructure.Configurations;
 using Infrastructure.DataSources;
 using Infrastructure.DbContexts;
 using Infrastructure.Services;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using MiniValidation;
 using Shared.DTO.Product.Output;
@@ -36,6 +37,7 @@ internal sealed class Create : IEndpoint
            })
            .WithTags("Products")
            .Produces<ICommandResult<ProductOutputDto?>>()
-           .WithName("Product.Create");
+           .WithName("Product.Create").RequireAuthorization(new AuthorizeAttribute { Roles = "Admin,Master" });
+
     }
 }
