@@ -1,7 +1,11 @@
 ﻿using Application.Controllers.Products;
 using Application.Interfaces.DataSources;
+using Application.Interfaces.Services;
+using Infrastructure.Configurations;
 using Infrastructure.DataSources;
 using Infrastructure.DbContexts;
+using Infrastructure.Services;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Shared.Result;
 
@@ -12,10 +16,12 @@ internal sealed class Delete : IEndpoint
     public void MapEndpoint(IEndpointRouteBuilder app)
     {
         app.MapDelete("api/products/{id}",
-           async (AppDbContext appDbContext, [FromRoute] Guid id) =>
+           async (AppDbContext appDbContext, FileStorageSettings _settings, [FromRoute] Guid id) =>
            {
+               IFileStorageService _fileStorage = new FileStorageService(_settings);
+
                IProductDataSource dataSource = new ProductDataSource(appDbContext);
-               ProductController _productController = new ProductController(dataSource);
+               ProductController _productController = new ProductController(dataSource,null,null, _fileStorage);
                var product = await _productController.DeleteProduct(id);
 
                return product.Succeeded ? Results.NoContent() : Results.BadRequest(product);
@@ -23,6 +29,7 @@ internal sealed class Delete : IEndpoint
            })
            .WithTags("Products")
            .Produces<ICommandResult>()
-           .WithName("Product.Delete");
+           .WithName("Product.Delete").RequireAuthorization(new AuthorizeAttribute { Roles = "Admin,Master" });
+
     }
 }
