@@ -26,7 +26,7 @@ namespace Application.Gateways
 
             return products.Select(x => new Product(x.Id, x.Name, x.Description, x.Price, x.CategorieId, x.CreatedAt,
              x.ProductIngredients.Select(k => new ProductIngredient(k.ProductId, k.IngredientId, k.Quantity)).ToList(),
-             x.ProductImages.Select(l => new ProductImage(l.Id, l.ProductId, l.Blob, l.Name, l.ImagePath, l.MimeType, l.FileName)).ToList(), new Domain.Entities.Categorie(x.Categorie.Name, x.Categorie.Id, x.Categorie.CreatedAt)
+             x.ProductImages.Select(l => new ProductImage(l.Id, l.ProductId, l.Blob, l.Name, l.ImagePath, l.MimeType, l.FileName)).ToList(), new Domain.Entities.Categorie(x.Categorie.Name, x.Categorie.Id, x.Categorie.CreatedAt, x.Categorie.IsEditavel)
              )).ToList();
         }
 
@@ -36,7 +36,7 @@ namespace Application.Gateways
 
             return product is not null ? new Product(product.Id, product.Name, product.Description, product.Price, product.CategorieId, product.CreatedAt,
                 product.ProductIngredients.Select(x => new ProductIngredient(x.ProductId, x.IngredientId, x.Quantity)).ToList(),
-                product.ProductImages.Select(x => new ProductImage(x.Id, x.ProductId, x.Blob, x.Name, x.ImagePath, x.MimeType, x.FileName)).ToList(), new Domain.Entities.Categorie(product.Categorie.Name, product.Categorie.Id, product.Categorie.CreatedAt)) : null;
+                product.ProductImages.Select(x => new ProductImage(x.Id, x.ProductId, x.Blob, x.Name, x.ImagePath, x.MimeType, x.FileName)).ToList(), new Domain.Entities.Categorie(product.Categorie.Name, product.Categorie.Id, product.Categorie.CreatedAt, product.Categorie.IsEditavel)) : null;
         }
 
         public async Task Delete(Guid id)
@@ -50,7 +50,7 @@ namespace Application.Gateways
 
             return products.Select(x => new Product(x.Id, x.Name, x.Description, x.Price, x.CategorieId, x.CreatedAt,
                x.ProductIngredients.Select(k => new ProductIngredient(k.ProductId, k.IngredientId, k.Quantity)).ToList(),
-               x.ProductImages.Select(l => new ProductImage(l.Id, l.ProductId, l.Blob, l.Name, l.ImagePath, l.MimeType, l.FileName)).ToList(), new Domain.Entities.Categorie(x.Categorie.Name, x.Categorie.Id, x.Categorie.CreatedAt)
+               x.ProductImages.Select(l => new ProductImage(l.Id, l.ProductId, l.Blob, l.Name, l.ImagePath, l.MimeType, l.FileName)).ToList(), new Domain.Entities.Categorie(x.Categorie.Name, x.Categorie.Id, x.Categorie.CreatedAt, x.Categorie.IsEditavel)
                )).ToList();
         }
  
@@ -58,7 +58,7 @@ namespace Application.Gateways
         {
             var productInput = new ProductInputDto(product.Id, product.CreatedAt, product.Name, product.Description, product.Price, product.CategorieId, 
                 product.ProductImages.Select(x => new ProductImageInputDto(x.Id, x.FileName, x.MimeType, x.ImagePath, x.Name, x.Blob, x.ProductId )).ToList() ,
-                product.ProductIngredients.Select(x => new ProductIngredientInputDto(x.IngredientId, x.Quantity, x.ProductId)).ToList(), new Shared.DTO.Categorie.Input.CategorieInputDto(product.Categorie.Id, product.Categorie.Name, product.Categorie.CreatedAt)
+                product.ProductIngredients.Select(x => new ProductIngredientInputDto(x.IngredientId, x.Quantity, x.ProductId)).ToList(), new Shared.DTO.Categorie.Input.CategorieInputDto(product.Categorie.Id, product.Categorie.Name, product.Categorie.IsEditavel, product.Categorie.CreatedAt)
               );            
 
             await _dataSource.Create(productInput);
@@ -68,7 +68,7 @@ namespace Application.Gateways
         {
             var productInput = new ProductInputDto(product.Id, product.CreatedAt, product.Name, product.Description, product.Price, product.CategorieId,
                 product.ProductImages.Select(x => new ProductImageInputDto(x.Id, x.FileName, x.MimeType, x.ImagePath, x.Name, x.Blob, x.ProductId)).ToList(),
-                product.ProductIngredients.Select(x => new ProductIngredientInputDto(x.IngredientId, x.Quantity, x.ProductId)).ToList(), new Shared.DTO.Categorie.Input.CategorieInputDto(product.Categorie.Id, product.Categorie.Name, product.Categorie.CreatedAt)
+                product.ProductIngredients.Select(x => new ProductIngredientInputDto(x.IngredientId, x.Quantity, x.ProductId)).ToList(), new Shared.DTO.Categorie.Input.CategorieInputDto(product.Categorie.Id, product.Categorie.Name, product.Categorie.IsEditavel, product.Categorie.CreatedAt)
               );
 
             await _dataSource.Update(productInput);
