@@ -1,4 +1,5 @@
-﻿using Shared.DTO.Categorie.Input;
+﻿using Domain.Entities;
+using Shared.DTO.Categorie.Input;
 
 namespace Application.Interfaces.DataSources
 {
@@ -7,5 +8,14 @@ namespace Application.Interfaces.DataSources
         Task<List<CategorieInputDto>> GetAllCategories();
 
         Task<CategorieInputDto?> GetCategorieById(Guid id);
+
+        Task CreateCategorie(CategorieInputDto categorie);
+
+        Task<CategorieInputDto?> GetByName(string name);
+
+        Task UpdateCategorie(CategorieInputDto categorie);
+
+        Task Delete(Guid id);
+
     }
 }
