@@ -4,17 +4,19 @@ namespace Domain.Entities
 {
     public class Categorie
     {
-        public Categorie(string name, Guid id, DateTime createdAt)
+        public Categorie(string name, Guid id, DateTime createdAt, bool isEditavel)
         {
             Name = name;
             Id = id;
             CreatedAt = createdAt;
+            IsEditavel = isEditavel;
         }
-        public Categorie(string name)
+        public Categorie(string name, bool isEditavel)
         {
             Name = name;
             Id = Guid.NewGuid();
             CreatedAt = DateTime.Now;
+            IsEditavel = isEditavel;
         }
 
         public DateTime CreatedAt { get; private set; }
@@ -23,6 +25,8 @@ namespace Domain.Entities
         public Guid Id { get; set; }
 
         public string Name { get; set; }
+
+        public bool IsEditavel { get; set; }
 
         public ICollection<Product> Produtos { get; set; } = new List<Product>();
 
