@@ -6,11 +6,12 @@ namespace Infrastructure.DbModels
     {
         public CategorieDbModel() { }
 
-        public CategorieDbModel(Guid id, string name, DateTime? createdAt = null )
+        public CategorieDbModel(Guid id, string name, bool isEditavel, DateTime? createdAt = null)
         {
             Name = name;
             Id = id;
             CreatedAt = createdAt ?? DateTime.Now;
+            IsEditavel = isEditavel;
         }
 
         public DateTime CreatedAt { get; private set; }
@@ -19,6 +20,7 @@ namespace Infrastructure.DbModels
 
         public string Name { get; set; }
 
+        public bool IsEditavel { get; set; } = false;
         public ICollection<ProductDbModel> Products { get; set; } = new List<ProductDbModel>();        
     }
 }
