@@ -1,5 +1,7 @@
-﻿using Infrastructure.DbContexts;
+﻿using Application.Interfaces.Services;
+using Infrastructure.DbContexts;
 using Infrastructure.DbModels;
+using Infrastructure.DbModels.UsersModelsAggregate;
 using Microsoft.EntityFrameworkCore;
 
 namespace Infrastructure.Persistence
@@ -8,39 +10,82 @@ namespace Infrastructure.Persistence
     {
 
         private readonly AppDbContext _context;
-        public DataSeeder(AppDbContext context)
+        private readonly IPasswordService _passwordService;
+        private bool _seedInDb = false;
+        public DataSeeder(AppDbContext context, IPasswordService passwordService)
         {
             _context = context;
+            _passwordService = passwordService;
         }
 
         public async Task Initialize()
         {
             await SeedIngredientes();
             await SeedCategories();
+            await SeedRoles();
+            await SeedUserAdm();
+            
+            if (_seedInDb) //Salva o contexto completo
+                await _context.SaveChangesAsync();
         }
 
+        private async Task SeedUserAdm()
+        {
+            string mailToAdm = "ale.alencarr@outlook.com.br";
+
+            var userDb = await _context.User.FirstOrDefaultAsync(x => x.Mail == mailToAdm);
+
+            if (userDb is null)
+            {
+                var userRoles = new List<UserRoleDbModel>
+                {
+                    new(Guid.Parse("00000000-0000-0000-0000-000000000001"), Guid.Parse("00000000-0000-0000-0000-000000000001"))
+                };
+
+                var userMock = new UserDbModel(Guid.Parse("00000000-0000-0000-0000-000000000001"), "Alexandre Alencar", mailToAdm, _passwordService.HashPassword("123465789"), userRoles, null, null);
+
+                await _context.User.AddAsync(userMock);
+                _seedInDb = true;
+            }
+        }
+        private async Task SeedRoles()
+        {
+            var rolesDb = await _context.Role.FirstOrDefaultAsync();
+
+            if (rolesDb is null)
+            {
+                var rolesMock = new List<RoleDbModel>()
+                {
+                    new RoleDbModel(Guid.Parse("00000000-0000-0000-0000-000000000001"), "Master"),
+                    new RoleDbModel(Guid.Parse("00000000-0000-0000-0000-000000000002"), "Admin"),
+                };
+
+                await _context.Role.AddRangeAsync(rolesMock);
+                _seedInDb = true;
+            }
+        }
         private async Task SeedCategories()
         {
-            var categoriesDb = await _context.Categories.FirstOrDefaultAsync();
+            var categoriesDb = await _context.Categorie.FirstOrDefaultAsync();
 
             if (categoriesDb is null)
             {
                 var categoriesMock = new List<CategorieDbModel>()
                 {
-                    new CategorieDbModel(Guid.Parse("00000000-0000-0000-0000-000000000001"), "Lanche"),
-                    new CategorieDbModel(Guid.Parse("00000000-0000-0000-0000-000000000002"), "Acompanhamento"),
-                    new CategorieDbModel(Guid.Parse("00000000-0000-0000-0000-000000000003"), "Bebida"),
-                    new CategorieDbModel(Guid.Parse("00000000-0000-0000-0000-000000000004"), "Sobremesa")
+                    new CategorieDbModel(Guid.Parse("00000000-0000-0000-0000-000000000001"), "Lanche",true),
+                    new CategorieDbModel(Guid.Parse("00000000-0000-0000-0000-000000000002"), "Acompanhamento",false),
+                    new CategorieDbModel(Guid.Parse("00000000-0000-0000-0000-000000000003"), "Bebida",false),
+                    new CategorieDbModel(Guid.Parse("00000000-0000-0000-0000-000000000004"), "Sobremesa",false)
                 };
 
-                await _context.Categories.AddRangeAsync(categoriesMock);
-                await _context.SaveChangesAsync();
+                await _context.Categorie.AddRangeAsync(categoriesMock);
+                _seedInDb = true;
             }
         }
 
         private async Task SeedIngredientes()
         {
-            var ingredientesDb = await _context.Ingredients.FirstOrDefaultAsync();
+            var ingredientesDb = await _context.Ingredient.FirstOrDefaultAsync();
 
             if (ingredientesDb is null)
             {
@@ -71,8 +116,8 @@ namespace Infrastructure.Persistence
                 new IngredientDbModel(Guid.Parse("10000000-0000-0000-0000-000000000024"), "Molho Trufado", 2.50m)
                 };
 
-                await _context.Ingredients.AddRangeAsync(ingredientesMock);
-                await _context.SaveChangesAsync();
+                await _context.Ingredient.AddRangeAsync(ingredientesMock);
+                _seedInDb = true;
             }
         }
     }
