@@ -1,4 +1,4 @@
 ﻿namespace Shared.DTO.Categorie.Input
 {
-    public record CategorieInputDto(Guid Id, string Name, DateTime CreatedAt);
+    public record CategorieInputDto(Guid Id, string Name, bool IsEditavel, DateTime CreatedAt);
 }
