@@ -1,4 +1,4 @@
 ﻿namespace Shared.DTO.Categorie.Output
 {
-    public record CategorieOutputDto(Guid Id, string Name);
+    public record CategorieOutputDto(Guid Id, string Name, bool IsEditavel);
 }
